+++ packages/adguard-api/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.4] - 2024-04-02
+
+### Changed
+- Update `@adguard/tswebextension` to `^1.0.22`.
+- Update `@adguard/filters-downloader` to `^2.2.0`.
+
 ## [2.1.3] - 2024-02-15
 
 ### Changed